+++ sdk/nodejs/ldap/authBackendUser.ts
@@ -13,7 +13,7 @@ import * as utilities from "../utilities";
  * import * as pulumi from "@pulumi/pulumi";
  * import * as vault from "@pulumi/vault";
  * 
- * const ldap = new vault.lDAP.AuthBackend("ldap", {
+ * const ldap = new vault.ldap.AuthBackend("ldap", {
  *     discoverdn: false,
  *     groupdn: "OU=Groups,DC=example,DC=org",
  *     groupfilter: "(&(objectClass=group)(member:1.2.840.113556.1.4.1941:={{.UserDN}}))",
@@ -23,7 +23,7 @@ import * as utilities from "../utilities";
  *     userattr: "sAMAccountName",
  *     userdn: "OU=Users,OU=Accounts,DC=example,DC=org",
  * });
- * const user = new vault.lDAP.AuthBackendUser("user", {
+ * const user = new vault.ldap.AuthBackendUser("user", {
  *     backend: ldap.path,
  *     policies: [
  *         "dba",
@@ -49,7 +49,7 @@ export class AuthBackendUser extends pulumi.CustomResource {
     }
 
     /** @internal */
-    public static readonly __pulumiType = 'vault:lDAP/authBackendUser:AuthBackendUser';
+    public static readonly __pulumiType = 'vault:ldap/authBackendUser:AuthBackendUser';
 
     /**
      * Returns true if the given object is an instance of AuthBackendUser.  This is designed to work even