+++ sdk/python/pulumi_vault/ldap/auth_backend_user.py
@@ -9,30 +9,35 @@ import pulumi.runtime
 from typing import Union
 from .. import utilities, tables
 
-class AuthBackendGroup(pulumi.CustomResource):
+class AuthBackendUser(pulumi.CustomResource):
     backend: pulumi.Output[str]
     """
     Path to the authentication backend
     """
-    groupname: pulumi.Output[str]
+    groups: pulumi.Output[list]
     """
-    The LDAP groupname
+    Override LDAP groups which should be granted to user
     """
     policies: pulumi.Output[list]
     """
-    Policies which should be granted to members of the group
+    Policies which should be granted to user
     """
-    def __init__(__self__, resource_name, opts=None, backend=None, groupname=None, policies=None, __props__=None, __name__=None, __opts__=None):
+    username: pulumi.Output[str]
+    """
+    The LDAP username
+    """
+    def __init__(__self__, resource_name, opts=None, backend=None, groups=None, policies=None, username=None, __props__=None, __name__=None, __opts__=None):
         """
-        Provides a resource to create a group in an [LDAP auth backend within Vault](https://www.vaultproject.io/docs/auth/ldap.html).
+        Provides a resource to create a user in an [LDAP auth backend within Vault](https://www.vaultproject.io/docs/auth/ldap.html).
         
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] backend: Path to the authentication backend
-        :param pulumi.Input[str] groupname: The LDAP groupname
-        :param pulumi.Input[list] policies: Policies which should be granted to members of the group
+        :param pulumi.Input[list] groups: Override LDAP groups which should be granted to user
+        :param pulumi.Input[list] policies: Policies which should be granted to user
+        :param pulumi.Input[str] username: The LDAP username
 
-        > This content is derived from https://github.com/terraform-providers/terraform-provider-vault/blob/master/website/docs/r/ldap_auth_backend_group.html.markdown.
+        > This content is derived from https://github.com/terraform-providers/terraform-provider-vault/blob/master/website/docs/r/ldap_auth_backend_user.html.markdown.
         """
         if __name__ is not None:
             warnings.warn("explicit use of __name__ is deprecated", DeprecationWarning)
@@ -52,38 +57,41 @@ class AuthBackendGroup(pulumi.CustomResource):
             __props__ = dict()
 
             __props__['backend'] = backend
-            if groupname is None:
-                raise TypeError("Missing required property 'groupname'")
-            __props__['groupname'] = groupname
+            __props__['groups'] = groups
             __props__['policies'] = policies
-        super(AuthBackendGroup, __self__).__init__(
-            'vault:lDAP/authBackendGroup:AuthBackendGroup',
+            if username is None:
+                raise TypeError("Missing required property 'username'")
+            __props__['username'] = username
+        super(AuthBackendUser, __self__).__init__(
+            'vault:ldap/authBackendUser:AuthBackendUser',
             resource_name,
             __props__,
             opts)
 
     @staticmethod
-    def get(resource_name, id, opts=None, backend=None, groupname=None, policies=None):
+    def get(resource_name, id, opts=None, backend=None, groups=None, policies=None, username=None):
         """
-        Get an existing AuthBackendGroup resource's state with the given name, id, and optional extra
+        Get an existing AuthBackendUser resource's state with the given name, id, and optional extra
         properties used to qualify the lookup.
         
         :param str resource_name: The unique name of the resulting resource.
         :param str id: The unique provider ID of the resource to lookup.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] backend: Path to the authentication backend
-        :param pulumi.Input[str] groupname: The LDAP groupname
-        :param pulumi.Input[list] policies: Policies which should be granted to members of the group
+        :param pulumi.Input[list] groups: Override LDAP groups which should be granted to user
+        :param pulumi.Input[list] policies: Policies which should be granted to user
+        :param pulumi.Input[str] username: The LDAP username
 
-        > This content is derived from https://github.com/terraform-providers/terraform-provider-vault/blob/master/website/docs/r/ldap_auth_backend_group.html.markdown.
+        > This content is derived from https://github.com/terraform-providers/terraform-provider-vault/blob/master/website/docs/r/ldap_auth_backend_user.html.markdown.
         """
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))
 
         __props__ = dict()
         __props__["backend"] = backend
-        __props__["groupname"] = groupname
+        __props__["groups"] = groups
         __props__["policies"] = policies
-        return AuthBackendGroup(resource_name, opts=opts, __props__=__props__)
+        __props__["username"] = username
+        return AuthBackendUser(resource_name, opts=opts, __props__=__props__)
     def translate_output_property(self, prop):
         return tables._CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop